+++ .prettierignore
@@ -7,3 +7,4 @@ docs
 node_modules
 CHANGELOG.md
 lib
+1e6.json
